requirements.txt
blinker==1.8.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
piexif==1.1.3
pillow==10.4.0
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.4
